Adding check & test_check
C0pied from on-going PR: https://github.com/roman-corgi/corgidrp/pull/105/files
