+++ corgidrp/calibrate_kgain.py
@@ -690,378 +690,6 @@ def calibrate_kgain(stack_arr, stack_arr2, emgain, min_val, max_val,
     
     return (kgain, mean_rn_gauss_e, mean_rn_std_e, ptc)
     
-
-
-"""
-Class to hold input-checking functions to minimize repetition
-Note: This module, used by calibrate_kgain.py and calibrate_nonlin.py is included
-here because at this moment it is not clear if the functions in the module are
-of general utility for corgidrp
-"""
-import numbers
-# String check support
-string_types = (str, bytes)
-# Int check support
-int_types = (int, np.integer)
-class check:
-    
-    class CheckException(Exception):
-        pass
-    
-    def _checkname(vname):
-        """
-        Internal check that we can use vname as a string for printing
-        """
-        if not isinstance(vname, string_types):
-            raise CheckException('vname must be a string when fed to check ' + \
-                                 'functions')
-        pass
-    
-    
-    def _checkexc(vexc):
-        """
-        Internal check that we can raise from the vexc object
-        """
-        if not isinstance(vexc, type): # pre-check it is class-like
-            raise CheckException('vexc must be a Exception, or an object ' + \
-                                 'descended from one when fed to check functions')
-        if not issubclass(vexc, Exception):
-            raise CheckException('vexc must be a Exception, or an object ' + \
-                                 'descended from one when fed to check functions')
-        pass
-    
-    
-    def real_positive_scalar(var, vname, vexc):
-        """
-        Checks whether an object is a real positive scalar.
-    
-        Arguments:
-         var: variable to check
-         vname: string to output in case of error for debugging
-         vexc: Exception to raise in case of error for debugging
-    
-        Returns:
-         returns var
-    
-        """
-        check._checkname(vname)
-        check._checkexc(vexc)
-    
-        if not isinstance(var, numbers.Number):
-            raise vexc(vname + ' must be scalar')
-        if not np.isrealobj(var):
-            raise vexc(vname + ' must be real')
-        if var <= 0:
-            raise vexc(vname + ' must be positive')
-        return var
-    
-    
-    def real_array(var, vname, vexc):
-        """
-        Checks whether an object is a real numpy array, or castable to one.
-    
-        Arguments:
-         var: variable to check
-         vname: string to output in case of error for debugging
-         vexc: Exception to raise in case of error for debugging
-    
-        Returns:
-         returns var
-    
-        """
-        check._checkname(vname)
-        check._checkexc(vexc)
-    
-        var = np.array(var, copy=False)  # cast to array
-        if len(var.shape) == 0:
-            raise vexc(vname + ' must have length > 0')
-        if not np.isrealobj(var):
-            raise vexc(vname + ' must be a real array')
-        # skip 'c' as we don't want complex; rest are non-numeric
-        if not var.dtype.kind in ['b', 'i', 'u', 'f']:
-            raise vexc(vname + ' must be a real numeric type to be real')
-        return var
-    
-    def oneD_array(var, vname, vexc):
-        """
-        Checks whether an object is a 1D numpy array, or castable to one.
-    
-        Arguments:
-         var: variable to check
-         vname: string to output in case of error for debugging
-         vexc: Exception to raise in case of error for debugging
-    
-        Returns:
-         returns var
-    
-        """
-        check._checkname(vname)
-        check._checkexc(vexc)
-    
-        var = np.array(var, copy=False) # cast to array
-        if len(var.shape) != 1:
-            raise vexc(vname + ' must be a 1D array')
-        if (not np.isrealobj(var)) and (not np.iscomplexobj(var)):
-            raise vexc(vname + ' must be a real or complex 1D array')
-        return var
-    
-    
-    def twoD_array(var, vname, vexc):
-        """
-        Checks whether an object is a 2D numpy array, or castable to one.
-    
-        Arguments:
-         var: variable to check
-         vname: string to output in case of error for debugging
-         vexc: Exception to raise in case of error for debugging
-    
-        Returns:
-         returns var
-    
-        """
-        check._checkname(vname)
-        check._checkexc(vexc)
-    
-        var = np.array(var, copy=False) # cast to array
-        if len(var.shape) != 2:
-            raise vexc(vname + ' must be a 2D array')
-        if (not np.isrealobj(var)) and (not np.iscomplexobj(var)):
-            raise vexc(vname + ' must be a real or complex 2D array')
-        return var
-    
-    
-    def twoD_square_array(var, vname, vexc):
-        """
-        Checks whether an object is a 2D square array_like.
-    
-        Arguments:
-         var: variable to check
-         vname: string to output in case of error for debugging
-         vexc: Exception to raise in case of error for debugging
-    
-        Returns:
-         returns var
-    
-        """
-        check._checkname(vname)
-        check._checkexc(vexc)
-    
-        var = np.array(var, copy=False) # cast to array
-        if len(var.shape) != 2:
-            raise vexc(vname + ' must be a 2D array')
-        else: # is 2-D
-            if not var.shape[0] == var.shape[1]:
-                raise vexc(vname + ' must be a square 2D array')
-        if (not np.isrealobj(var)) and (not np.iscomplexobj(var)):
-            raise vexc(vname + ' must be a real or complex square 2D array')
-        return var
-    
-    def threeD_array(var, vname, vexc):
-        """
-        Checks whether an object is a 3D numpy array, or castable to one.
-    
-        Arguments:
-         var: variable to check
-         vname: string to output in case of error for debugging
-         vexc: Exception to raise in case of error for debugging
-    
-        Returns:
-         returns var
-    
-        """
-        check._checkname(vname)
-        check._checkexc(vexc)
-    
-        var = np.array(var, copy=False) # cast to array
-        if len(var.shape) != 3:
-            raise vexc(vname + ' must be a 3D array')
-        if (not np.isrealobj(var)) and (not np.iscomplexobj(var)):
-            raise vexc(vname + ' must be a real or complex 3D array')
-        return var
-    
-    
-    
-    def real_scalar(var, vname, vexc):
-        """
-        Checks whether an object is a real scalar.
-    
-        Arguments:
-         var: variable to check
-         vname: string to output in case of error for debugging
-         vexc: Exception to raise in case of error for debugging
-    
-        Returns:
-         returns var
-    
-        """
-        check._checkname(vname)
-        check._checkexc(vexc)
-    
-        if not isinstance(var, numbers.Number):
-            raise vexc(vname + ' must be scalar')
-        if not np.isrealobj(var):
-            raise vexc(vname + ' must be real')
-        return var
-    
-    
-    def real_nonnegative_scalar(var, vname, vexc):
-        """
-        Checks whether an object is a real nonnegative scalar.
-    
-        Arguments:
-         var: variable to check
-         vname: string to output in case of error for debugging
-         vexc: Exception to raise in case of error for debugging
-    
-        Returns:
-         returns var
-    
-        """
-        check._checkname(vname)
-        check._checkexc(vexc)
-    
-        if not isinstance(var, numbers.Number):
-            raise vexc(vname + ' must be scalar')
-        if not np.isrealobj(var):
-            raise vexc(vname + ' must be real')
-        if var < 0:
-            raise vexc(vname + ' must be nonnegative')
-        return var
-    
-    def positive_scalar_integer(var, vname, vexc):
-        """
-        Checks whether an object is a positive scalar integer.
-    
-        Arguments:
-         var: variable to check
-         vname: string to output in case of error for debugging
-         vexc: Exception to raise in case of error for debugging
-    
-        Returns:
-         returns var
-    
-        """
-        check._checkname(vname)
-        check._checkexc(vexc)
-    
-        if not isinstance(var, numbers.Number):
-            raise vexc(vname + ' must be scalar')
-        if not isinstance(var, int_types):
-            raise vexc(vname + ' must be integer')
-        if var <= 0:
-            raise vexc(vname + ' must be positive')
-        return var
-    
-    
-    def nonnegative_scalar_integer(var, vname, vexc):
-        """
-        Checks whether an object is a nonnegative scalar integer.
-    
-        Arguments:
-         var: variable to check
-         vname: string to output in case of error for debugging
-         vexc: Exception to raise in case of error for debugging
-    
-        Returns:
-         returns var
-    
-        """
-        check._checkname(vname)
-        check._checkexc(vexc)
-    
-        if not isinstance(var, numbers.Number):
-            raise vexc(vname + ' must be scalar')
-        if not isinstance(var, int_types):
-            raise vexc(vname + ' must be integer')
-        if var < 0:
-            raise vexc(vname + ' must be nonnegative')
-        return var
-    
-    
-    def scalar_integer(var, vname, vexc):
-        """
-        Checks whether an object is a scalar integer (no sign dependence).
-    
-        Arguments:
-         var: variable to check
-         vname: string to output in case of error for debugging
-         vexc: Exception to raise in case of error for debugging
-    
-        Returns:
-         returns var
-    
-        """
-        check._checkname(vname)
-        check._checkexc(vexc)
-    
-        if not isinstance(var, numbers.Number):
-            raise vexc(vname + ' must be scalar')
-        if not isinstance(var, int_types):
-            raise vexc(vname + ' must be integer')
-        return var
-    
-    
-    def string(var, vname, vexc):
-        """
-        Checks whether an object is a string.
-    
-        Arguments:
-         var: variable to check
-         vname: string to output in case of error for debugging
-         vexc: Exception to raise in case of error for debugging
-    
-        Returns:
-         returns var
-    
-        """
-        check._checkname(vname)
-        check._checkexc(vexc)
-    
-        if not isinstance(var, string_types):
-            raise vexc(vname + ' must be a string')
-        return var
-    
-    def boolean(var, vname, vexc):
-        """
-        Checks whether an object is a bool.
-    
-        Arguments:
-         var: variable to check
-         vname: string to output in case of error for debugging
-         vexc: Exception to raise in case of error for debugging
-    
-        Returns:
-         returns var
-    
-        """
-        check._checkname(vname)
-        check._checkexc(vexc)
-    
-        if not isinstance(var, bool):
-            raise vexc(vname + ' must be a bool')
-        return var
-    
-    
-    def dictionary(var, vname, vexc):
-        """
-        Checks whether an object is a dictionary.
-    
-        Arguments:
-         var: variable to check
-         vname: string to output in case of error for debugging
-         vexc: Exception to raise in case of error for debugging
-    
-        Returns:
-         returns var
-    
-        """
-        check._checkname(vname)
-        check._checkexc(vexc)
-    
-        if not isinstance(var, dict):
-            raise vexc(vname + ' must be a dict')
-        return var
-
 def copy_and_cast(stack_arr_drp, stack_arr2_drp):
     """ Copies and casts input stacks of CORGIDRP Data Image objects into numpy
     arrays to perform computations.
